:fire: removing, refacoring and moving files.
