+++ test/domain/usecases/read_anime_details_test.dart
@@ -3,10 +3,11 @@ import 'package:anitube_crawler_api/src/domain/entities/parser/ihtml_parser.dart
 import 'package:anitube_crawler_api/src/domain/irepository/ianime_details_repository.dart';
 import 'package:anitube_crawler_api/src/domain/usecases/read_anime_details.dart';
 import 'package:anitube_crawler_api/src/external/anitube/parser/anitube_anime_details_parser.dart';
-import 'package:anitube_crawler_api/test_resources/anime_details_page.dart';
 import 'package:test/test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:anitube_crawler_api/src/domain/entities/AnimeDetails.dart';
+
+import '../../test_resources/anime_details_page.dart';
 
 class MockedRepository extends Mock implements IAnimeDetailsRepository {}
