+++ test/external/parser/anitube_episode_details_parser_test.dart
@@ -1,8 +1,8 @@
 import 'package:anitube_crawler_api/src/domain/entities/EpisodeDetails.dart';
 import 'package:anitube_crawler_api/src/domain/exceptions/CrawlerApiException.dart';
 import 'package:anitube_crawler_api/src/external/anitube/parser/anitube_episode_details_page_parser.dart';
-import 'package:anitube_crawler_api/test_resources/episode_details_page.dart';
 import 'package:test/test.dart';
+import '../../test_resources/episode_details_page.dart';
 
 main() {
   AnitubeEpisodeDetailsPageParser parser;
